pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "betha_extractor"
description = "Ferramenta CLI para extração de dados da API Betha"
version = "0.1.0"
authors = [
    { name = "Lucas" }
]
requires-python = ">=3.10"
dependencies = [
    "requests>=2.32.3",
    "python-dotenv>=1.0.1",
    "urllib3>=2.2.2",
    "rich>=13.7.1",
    "typer>=0.12.3",
]

[project.scripts]
betha-extractor = "betha_extractor.main:app"

[tool.hatch.build.targets.wheel]
packages = ["betha_extractor"]
